[FE][Fix][Refactor] : InputBox 디렉토리 이동
